+++ views/modal/edit_quiz.php
@@ -172,7 +172,7 @@ if ( ! $quiz){
                             <div class="tutor-quiz-builder-col">
                                 <?php
                                 $default_attempts_allowed = tutor_utils()->get_option('quiz_attempts_allowed');
-                                $attempts_allowed = tutor_utils()->get_quiz_option($quiz_id, 'attempts_allowed', $default_attempts_allowed);
+                                $attempts_allowed = (int) tutor_utils()->get_quiz_option($quiz_id, 'attempts_allowed', $default_attempts_allowed);
                                 ?>
 
                                 <div class="tutor-field-type-slider" data-min="0" data-max="20">
